Sometimes a null DiscordMessage is passed as a parameter and causes a crash, so the method now checks for null.

+++ DiscordBot.cs
@@ -72,7 +72,8 @@ namespace DiscordSharpTest
 
         virtual public void DeleteMessage(DiscordMessage target)
         {
-            Client.DeleteMessage(target);
+            if (target != null)
+                Client.DeleteMessage(target);
         }
 
         virtual public DiscordMessage GetMessageByID(string messageID, DiscordChannel channel)
